setup gemm

+++ src/cuda/gemm_tensorcores/src/setup_gemm.cu
@@ -263,7 +263,7 @@ void setup_gemm_dmr(Log& log) {
 
 	if (log.precision == "double") {
 		if (log.dmr == "mixed") {
-			DMRMixedGemmCaller<8, double, float> gemm_obj(log.size_matrices,
+			DMRMixedGemmCaller<8, float, double> gemm_obj(log.size_matrices,
 					log.size_matrices);
 			setup_execute(log, gemm_obj, THRESHOLD_1);
 		} else {
